Style: 컴포넌트 분리 및 리뷰에 맞게 수정(버튼, 디바운스빼고)
