+++ src/components/AddRhythm.tsx
@@ -1,17 +1,14 @@
 import React, { useState } from 'react';
 import ButtonComponent from './ui/ButtonComponent';
 import Loading from './ui/Loading';
-import { addOrUpdateNewRhythm } from '../api/firebase';
 import { useGetUser } from '../store/useUserStore';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 import styled from 'styled-components';
-import { color, darkTheme, lightTheme } from '../css/styles.theme';
-import { DatePicker, Space, TimePicker } from 'antd';
 import dayjs from 'dayjs';
-import { FcCalendar, FcAlarmClock } from 'react-icons/fc';
-import { PiPencilCircleFill } from 'react-icons/pi';
-import { FaCheck } from 'react-icons/fa';
-import { highlighter } from '../css/styles.highlighter';
+import SelectIcon from './SelectIcon';
+import RhythmTitleInput from './RhythmTitleInput';
+import { useAddRhythm } from '../hooks/useAddRhythm';
+import TimeAndPeriod from './TimeAndPeriod';
+import SelectHighlighter from './SelectHighlighter';
 export interface RhythmItem {
   id: string;
   time: string;
@@ -38,38 +35,14 @@ interface CloseModalProps {
   onClick: () => void;
 }
 
-const ICON_OPTIONS = ['✅', '💊', '💪', '📖', '🔥'];
-const FORMAT = 'HH:mm';
-
 export default function AddRhythm({ onClick }: CloseModalProps) {
   const [rhythm, setRhythm] = useState(initialRhythm);
   const [isLoading, setIsLoading] = useState(false);
   const [selectedIcon, setSelectedIcon] = useState('✅');
-  const [isVisible, setIsVisible] = useState(false);
   const [selectedBackgroundColor, setSelectedBackgroundColor] = useState('');
   const user = useGetUser();
   const uid = user.uid;
-  const queryClient = useQueryClient();
-  const { RangePicker } = DatePicker;
-
-  const addRhythm = useMutation<
-    void,
-    Error,
-    { uid: string; rhythm: RhythmItem }
-  >({
-    mutationFn: ({ uid, rhythm }) => addOrUpdateNewRhythm(uid, rhythm),
-    onError: (error) => {
-      console.error(error);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['rhythms', uid] });
-    },
-  });
-
-  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
-    const { name, value } = e.target;
-    setRhythm((rhythm) => ({ ...rhythm, [name]: value }));
-  }
+  const addRhythm = useAddRhythm(uid);
 
   async function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
@@ -94,15 +67,11 @@ export default function AddRhythm({ onClick }: CloseModalProps) {
     }
   }
 
-  const handleVisibleClick = () => {
-    setIsVisible(!isVisible);
-  };
+  function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
+    const { name, value } = e.target;
+    setRhythm((rhythm) => ({ ...rhythm, [name]: value }));
+  }
 
-  const handleIconSelect = (icon: string) => {
-    setSelectedIcon(icon);
-    setRhythm({ ...rhythm, icon });
-    setIsVisible(false);
-  };
   const handleColorSelect = (backgroundColor: string) => {
     if (selectedBackgroundColor === backgroundColor) {
       setSelectedBackgroundColor('');
@@ -140,99 +109,25 @@ export default function AddRhythm({ onClick }: CloseModalProps) {
         />
       </StyledAddRhythmHead>
       <StyledAddRhythmForm onSubmit={handleSubmit}>
-        {/* // 진짜 너무어렵네ㅔㅔㅔㅔㅔ
-            버튼 클릭하면 아이콘 선택창 조그맣게 만들고
-            버튼을 만들어서 대표아이콘 처럼 보익 해보자       
-        */}
-        <StyledSelectIconBox>
-          <StyledSelectIconButton type='button' onClick={handleVisibleClick}>
-            {selectedIcon}
-          </StyledSelectIconButton>
-          {isVisible && (
-            <div>
-              <StyledIconOptionWrapper>
-                {ICON_OPTIONS.map((icon) => (
-                  <StyledIconOption
-                    key={icon}
-                    onClick={() => handleIconSelect(icon)}
-                    $isSelected={icon === selectedIcon}
-                  >
-                    {icon}
-                  </StyledIconOption>
-                ))}
-              </StyledIconOptionWrapper>
-            </div>
-          )}
-        </StyledSelectIconBox>
-
-        <StyledAddRhythmTextInput
-          type='text'
-          name='title'
-          value={rhythm.title ?? ''}
-          placeholder='Title'
-          required
-          onChange={handleChange}
+        <SelectIcon
+          selectedIcon={selectedIcon}
+          onSelect={(icon) => {
+            setSelectedIcon(icon);
+            setRhythm({ ...rhythm, icon });
+          }}
         />
-        <StyledAddRhythmTimeAndPeriodBox>
-          <div>
-            <StyledAddRhythmTitle>
-              <StyledAddRhythmIcon $fontSize={'1.3rem'}>
-                <FcCalendar />
-              </StyledAddRhythmIcon>
-              Period
-            </StyledAddRhythmTitle>
-            <Space direction='vertical'>
-              <RangePicker
-                value={[
-                  rhythm.startDate ? dayjs(rhythm.startDate) : null,
-                  rhythm.endDate ? dayjs(rhythm.endDate) : null,
-                ]}
-                onChange={handleRangeChange}
-              />
-            </Space>
-          </div>
-          <div>
-            <StyledAddRhythmTitle>
-              <StyledAddRhythmIcon $fontSize={'1.2rem'}>
-                <FcAlarmClock />
-              </StyledAddRhythmIcon>
-              Time
-            </StyledAddRhythmTitle>
-            <TimePicker
-              value={rhythm.time ? dayjs(rhythm.time, FORMAT) : null}
-              format={FORMAT}
-              onChange={handleTimeChange}
-            />
-          </div>
-        </StyledAddRhythmTimeAndPeriodBox>
-        <div>
-          <StyledAddRhythmTitle>
-            <StyledAddRhythmIcon $fontSize={'1.4rem'}>
-              <PiPencilCircleFill />
-            </StyledAddRhythmIcon>
-            Highlighter
-          </StyledAddRhythmTitle>
-          <StyledAddRhythmColorWrapper>
-            {Object.keys(highlighter).map(
-              (colorKey: keyof typeof highlighter) => (
-                <StyledAddRhythmColor
-                  key={colorKey}
-                  onClick={() => handleColorSelect(highlighter[colorKey])}
-                  $color={highlighter[colorKey]}
-                  $isSelected={
-                    selectedBackgroundColor === highlighter[colorKey]
-                  }
-                >
-                  {selectedBackgroundColor === highlighter[colorKey] ? (
-                    <FaCheck />
-                  ) : (
-                    ''
-                  )}
-                </StyledAddRhythmColor>
-              )
-            )}
-          </StyledAddRhythmColorWrapper>
-        </div>
+        <RhythmTitleInput title={rhythm.title} onChange={handleChange} />
+        <TimeAndPeriod
+          time={rhythm.time}
+          startDate={rhythm.startDate}
+          endDate={rhythm.endDate}
+          onTimeChange={handleTimeChange}
+          onRangeChange={handleRangeChange}
+        />
+        <SelectHighlighter
+          selectedBackgroundColor={selectedBackgroundColor}
+          onColorSelect={handleColorSelect}
+        />
 
         <ButtonComponent text={isLoading ? <Loading /> : 'rhythm 추가'} />
       </StyledAddRhythmForm>
@@ -253,118 +148,4 @@ const StyledAddRhythmForm = styled.form`
   flex-direction: column;
   justify-content: space-between;
   gap: 1rem;
-`;
-
-const StyledAddRhythmTextInput = styled.input`
-  width: 97%;
-  padding: 0.5rem;
-  border: none;
-  background-color: ${color.gray};
-  font-size: 1.1rem;
-`;
-
-// Icon
-
-const StyledSelectIconBox = styled.div`
-  display: flex;
-  align-items: center;
-`;
-const StyledIconOptionWrapper = styled.div`
-  display: flex;
-  position: relative;
-  background: ${color.lightGray};
-  border-radius: 1rem;
-  font-size: 1.3rem;
-
-  &:after {
-    content: '';
-    position: absolute;
-    left: 0;
-    top: 50%;
-    width: 0;
-    height: 0;
-    border: 11px solid transparent;
-    border-right-color: ${color.lightGray};
-    border-left: 0;
-    margin-top: -11px;
-    margin-left: -10px;
-  }
-`;
-
-const StyledIconOption = styled.div<{ $isSelected: boolean }>`
-  padding: 0.7rem;
-  cursor: pointer;
-  background-color: ${({ $isSelected }) =>
-    $isSelected ? lightTheme.accentColor : 'transparent'};
-  &:hover {
-    background-color: ${lightTheme.accentColor};
-  }
-  &:first-child {
-    border-radius: 1rem 0rem 0rem 1rem;
-    &:hover {
-      background-color: ${lightTheme.accentColor};
-    }
-  }
-
-  &:last-child {
-    border-radius: 0rem 1rem 1rem 0rem;
-    &:hover {
-      background-color: ${lightTheme.accentColor};
-    }
-  }
-`;
-
-const StyledSelectIconButton = styled.button`
-  border-radius: 100%;
-  border: none;
-  font-size: 2rem;
-  width: 4.5rem;
-  height: 4.5rem;
-  margin-right: 1rem;
-  background-color: transparent;
-  border: 3px solid ${darkTheme.primaryColor};
-`;
-
-// time & period
-
-const StyledAddRhythmTimeAndPeriodBox = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-const StyledAddRhythmIcon = styled.div<{ $fontSize: string }>`
-  font-size: ${({ $fontSize }) => $fontSize};
-  display: flex;
-  align-items: center;
-  margin-right: 0.2rem;
-`;
-const StyledAddRhythmTitle = styled.h2`
-  display: flex;
-  align-items: center;
-  margin: 0.5rem 0rem 0.7rem 0rem;
-  font-size: 1.1rem;
-`;
-
-// backgroundColor
-const StyledAddRhythmColorWrapper = styled.div`
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  gap: 1rem;
-`;
-
-const StyledAddRhythmColor = styled.div<{
-  $color: string;
-  $isSelected: boolean;
-}>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: ${({ $color }) => $color};
-  width: 2.4rem;
-  height: 2.4rem;
-  border-radius: 100%;
-  border: 2px solid;
-  color: ${darkTheme.errorColor};
-  opacity: ${({ $isSelected }) => ($isSelected ? '0.8' : '1')};
-  border-color: ${({ $isSelected }) =>
-    $isSelected ? darkTheme.secondaryColor : lightTheme.placeholderColor};
 `;
